+++ internal/client/runner/allocrunner/alloc_runner_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/umbracle/vesta/internal/client/runner/docker"
 	"github.com/umbracle/vesta/internal/client/runner/mock"
-	"github.com/umbracle/vesta/internal/client/runner/proto"
 	"github.com/umbracle/vesta/internal/client/runner/state"
+	proto "github.com/umbracle/vesta/internal/client/runner/structs"
 	"github.com/umbracle/vesta/internal/testutil"
 )
