+++ internal/client/runner/hooks/hooks.go
@@ -2,7 +2,7 @@ package hooks
 
 import (
 	"github.com/hashicorp/go-hclog"
-	"github.com/umbracle/vesta/internal/client/runner/proto"
+	proto "github.com/umbracle/vesta/internal/client/runner/structs"
 )
 
 type TaskHook interface {
