+++ internal/client/runner/runner.go
@@ -7,8 +7,8 @@ import (
 	"github.com/umbracle/vesta/internal/client/runner/allocrunner"
 	"github.com/umbracle/vesta/internal/client/runner/docker"
 	"github.com/umbracle/vesta/internal/client/runner/hooks"
-	"github.com/umbracle/vesta/internal/client/runner/proto"
 	"github.com/umbracle/vesta/internal/client/runner/state"
+	proto "github.com/umbracle/vesta/internal/client/runner/structs"
 )
 
 type Config struct {
