+++ internal/client/runner/runner_test.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/umbracle/vesta/internal/client/runner/proto"
 	"github.com/umbracle/vesta/internal/client/runner/state"
+	proto "github.com/umbracle/vesta/internal/client/runner/structs"
 	"github.com/umbracle/vesta/internal/uuid"
 )
